+++ src/lib/types/leaderboard.ts
@@ -3,6 +3,7 @@ export interface LeaderboardEntry {
     lastUpdated: number;
     level: number;
     picture?: string;
+    rank: number;
     self?: boolean;
     userId?: string;
     username: string;